// src/@types/data.d.ts
export type nodeProps = {
  id: string,
  name: string,
  img: any,
  group: string,
  address: string,
  designation: string,
  designationDescription: string,
  university: {
    schoolName: string,
    degree: string,
    startDate: string,
    endDate: string,
  },
  noOfPeers: number,
  noOfFollowers: number,
  patientsServed: number,
  successRate: number,
  __indexColor: string,
  index: number,
  x: number,
  y: number,
  vx: number,
  vy: number,
  fx: number,
  fy: number,
};